+++ lib/shared/src/telemetry-v2/TelemetryRecorderProvider.ts
@@ -15,8 +15,7 @@ import {
     type ClientConfigurationWithAccessToken,
     type CodyIDE,
 } from '../configuration'
-import { SourcegraphGraphQLAPIClient } from '../sourcegraph-api/graphql'
-import type { LogEventMode } from '../sourcegraph-api/graphql/client'
+import { type LogEventMode, graphqlClient } from '../sourcegraph-api/graphql/client'
 import { GraphQLTelemetryExporter } from '../sourcegraph-api/telemetry/GraphQLTelemetryExporter'
 import { MockServerTelemetryExporter } from '../sourcegraph-api/telemetry/MockServerTelemetryExporter'
 
@@ -69,7 +68,7 @@ export class TelemetryRecorderProvider extends BaseTelemetryRecorderProvider<
         anonymousUserID: string,
         legacyBackcompatLogEventMode: LogEventMode
     ) {
-        const client = new SourcegraphGraphQLAPIClient(config)
+        graphqlClient.setConfig(config)
         const clientName = extensionDetails.telemetryClientName
             ? extensionDetails.telemetryClientName
             : `${extensionDetails.ide || 'unknown'}.Cody`
@@ -81,7 +80,7 @@ export class TelemetryRecorderProvider extends BaseTelemetryRecorderProvider<
             },
             process.env.CODY_TELEMETRY_EXPORTER === 'testing'
                 ? TESTING_TELEMETRY_EXPORTER.withAnonymousUserID(anonymousUserID)
-                : new GraphQLTelemetryExporter(client, anonymousUserID, legacyBackcompatLogEventMode),
+                : new GraphQLTelemetryExporter(anonymousUserID, legacyBackcompatLogEventMode),
             [
                 new ConfigurationMetadataProcessor(config, authStatusProvider),
                 // Generate timestamps when recording events, instead of serverside